backend/chatbot/requirements.txt
fastapi==0.110.2
httpx==0.28.1
aio-pika==9.5.5
python-dotenv==1.0.1
tenacity==8.2.3
langchain-openai
langchain-core
langchain-community
langchain
pydantic==2.7.4
uvicorn==0.29.0
typing-extensions==4.11.0
langchain-redis
